简化日志路径

+++ whimbox/common/logger.py
@@ -4,6 +4,9 @@ import os, sys
 import time
 from whimbox.common.path_lib import *
 from whimbox.common.cvars import DEBUG_MODE
+import datetime
+
+log_file_prefix = 'whimbox-'
 
 def get_logger_format_time():
     return time.strftime('%Y-%m-%d %H:%M:%S', time.localtime())
@@ -21,18 +24,17 @@ def warning_once(self, message):
 def demo(self, message):
     self.info(f"DEMO: {message}")
 
-import datetime
-
 def delete_files(path, days):
     now = datetime.datetime.now()
     for root, dirs, files in os.walk(path):
         for file in files:
-            file_path = os.path.join(root, file)
-            modified_time = datetime.datetime.fromtimestamp(os.path.getmtime(file_path))
-            if (now - modified_time).days > days:
-                if DEBUG_MODE:
-                    print(f"Log File Delete: Deleting file {file_path} Last modified {modified_time} Days since modified {(now - modified_time).days} Days to delete {days}")
-                os.remove(file_path)
+            if file.startswith(log_file_prefix):
+                file_path = os.path.join(root, file)
+                modified_time = datetime.datetime.fromtimestamp(os.path.getmtime(file_path))
+                if (now - modified_time).days > days:
+                    if DEBUG_MODE:
+                        print(f"Log File Delete: Deleting file {file_path} Last modified {modified_time} Days since modified {(now - modified_time).days} Days to delete {days}")
+                    os.remove(file_path)
 
 delete_files(LOG_PATH, 7)
 
@@ -41,7 +43,7 @@ delete_files(LOG_PATH, 7)
 logger.remove(handler_id=None)
 logger.warning_once = types.MethodType(warning_once, logger)
 logger.demo = types.MethodType(demo, logger)
-logger.add(os.path.join(LOG_PATH, "{time:YYYY-MM-DD}/{time:YYYY-MM-DD}.log"), level="TRACE", backtrace=True)
+logger.add(os.path.join(LOG_PATH, log_file_prefix + "{time:YYYY-MM-DD}.log"), level="TRACE", backtrace=True)
 if DEBUG_MODE:
     logger.add(sys.stdout, level="TRACE", backtrace=True)
 else:
